Upgrading to syn 0.12 and quote 0.4
